+++ Example/BitmovinYospaceModule_Example_tvOS/ViewController.swift
@@ -18,19 +18,19 @@ class ViewController: UIViewController {
     @IBOutlet var adLabel: UILabel!
     
     lazy var playerView: PlayerView = {
-        let playerView = PlayerView(player: player, frame: .zero)
+        let playerView = PlayerView(player: player.bitmovinPlayer(), frame: .zero)
         playerView.autoresizingMask = [.flexibleHeight, .flexibleWidth]
         playerView.frame = playerContainer.bounds
         return playerView
     }()
     
     lazy var player: BitmovinYospacePlayer = {
-        let configuration = PlayerConfiguration()
-        configuration.playbackConfiguration.isAutoplayEnabled = true
+        let playerConfig = PlayerConfig()
+        playerConfig.playbackConfig.isAutoplayEnabled = true
 
         let player = BitmovinYospacePlayer(
-            configuration: configuration,
-            yospaceConfiguration: .init()
+            playerConfig: playerConfig,
+            yospaceConfig: .init()
         )
 
         player.add(listener: self)
@@ -52,11 +52,10 @@ class ViewController: UIViewController {
             return
         }
 
-        let sourceConfig = SourceConfiguration()
-        sourceConfig.addSourceItem(item: SourceItem(hlsSource: HLSSource(url: streamUrl)))
-        let config = YospaceSourceConfiguration(yospaceAssetType: .linear)
+        let sourceConfig = SourceConfig(url: streamUrl, type: .hls)
+        let yospaceSourceConfig = YospaceSourceConfig(yospaceAssetType: .linear)
 
-        player.load(sourceConfiguration: sourceConfig, yospaceSourceConfiguration: config)
+        player.load(sourceConfig: sourceConfig, yospaceSourceConfig: yospaceSourceConfig)
     }
 
     @IBAction func vodButtonClicked(sender: UIButton) {
@@ -64,16 +63,15 @@ class ViewController: UIViewController {
             return
         }
 
-        let sourceConfig = SourceConfiguration()
-        sourceConfig.addSourceItem(item: SourceItem(hlsSource: HLSSource(url: streamUrl)))
-        let config = YospaceSourceConfiguration(yospaceAssetType: .vod)
+        let sourceConfig = SourceConfig(url: streamUrl, type: .hls)
+        let yospaceSourceConfig = YospaceSourceConfig(yospaceAssetType: .vod)
 
-        player.load(sourceConfiguration: sourceConfig, yospaceSourceConfiguration: config)
+        player.load(sourceConfig: sourceConfig, yospaceSourceConfig: yospaceSourceConfig)
     }
 }
 
 extension ViewController: PlayerListener {
-    func onTimeChanged(_ event: TimeChangedEvent) {
-        adLabel.text = "Ad: \(player.isAd) time=\(Double(round(10*player.currentTime)/10))"
+    func onTimeChanged(_ event: TimeChangedEvent, player: Player) {
+        adLabel.text = "Ad: \(self.player.isAd()) time=\(Double(round(10*self.player.currentTime())/10))"
     }
 }